refactor: rename `http_server` folder to make it consistent with other ones
